11.20.1 Updates: Added files to debug Added “current day” as a global variable, then created a function and event listener to call the current date and time.
11.21.1 Updates:
Met with study group to start the “Save” event listener using jquery but struggled to complete

11.21.2 Updates:
Met with instructor during office hours to define the parent and description variables using classes assigned in HTML and store saved items locally.
Then differentiated between past, present and future time slots using the “.time-block” class and an if statement written with dayjs.

11.24.1 Updates:
Tested the scheduler and found that saved items weren’t remaining on display after clicking “Refresh”
Stepped away from the computer to sleep on it.

11.25 Updates:
Tested again to find it wasn’t an issue for items saved to hour 9.
Looked in the event listener and found “#hour-9” within a getItem method.
Copy + pasted that same line to “get” items for remaining hours.
Added learnings and insights to ReadMe file